+++ src/dmi.rs
@@ -2,7 +2,7 @@ use crate::{
     commands::DmiOp,
     device::WchLink,
     error::{AbstractcsCmdErr, Error, Result},
-    regs::{Abstractcs, DMReg, Dmstatus},
+    regs::{Abstractcs, DMReg, Dmcontrol, Dmstatus},
 };
 use std::{thread, time::Duration};
 
@@ -116,6 +116,19 @@ impl<'a, D: DebugModuleInterface> Algorigthm<'a, D> {
         Ok(())
     }
 
+    pub fn reset_debug_module(&mut self) -> Result<()> {
+        self.dmi.dmi_write(0x10, 0x00000000)?;
+        self.dmi.dmi_write(0x10, 0x00000001)?;
+
+        let dmcontrol = self.dmi.read_dmi_reg::<Dmcontrol>()?;
+
+        if dmcontrol.dmactive() {
+            Ok(())
+        } else {
+            Err(Error::DmiFailed)
+        }
+    }
+
     pub fn read_mem32(&mut self, addr: u32) -> Result<u32> {
         self.dmi.dmi_write(0x20, 0x0002a303)?; // lw x6,0(x5)
         self.dmi.dmi_write(0x21, 0x00100073)?; // ebreak
@@ -142,7 +155,6 @@ impl<'a, D: DebugModuleInterface> Algorigthm<'a, D> {
         }
 
         let data0 = self.dmi.dmi_read(0x04)?;
-
         Ok(data0)
     }
 
@@ -182,7 +194,39 @@ impl<'a, D: DebugModuleInterface> Algorigthm<'a, D> {
         Ok(())
     }
 
-    pub fn unlock_flash(&mut self) {}
+    pub fn write_mem8(&mut self, addr: u32, data: u8) -> Result<()> {
+        self.dmi.dmi_write(0x20, 0x00728023)?; // sb x7,0(x5)
+        self.dmi.dmi_write(0x21, 0x00100073)?; // ebreak
+
+        self.dmi.dmi_write(0x04, addr)?; // data0 <- address
+
+        self.clear_abstractcs_cmderr()?;
+        self.dmi.dmi_write(0x17, 0x00231005)?; // x5 <- data0
+
+        let abstractcs: Abstractcs = self.dmi.read_dmi_reg()?;
+        log::trace!("{:?}", abstractcs);
+        if abstractcs.busy() {
+            return Err(Error::AbstractCommandError(AbstractcsCmdErr::Busy)); //resue busy
+        }
+        if abstractcs.cmderr() != 0 {
+            AbstractcsCmdErr::try_from_cmderr(abstractcs.cmderr() as _)?;
+        }
+
+        self.dmi.dmi_write(0x04, data as u32)?; // data0 <- data
+        self.clear_abstractcs_cmderr()?;
+
+        self.dmi.dmi_write(0x17, 0x00271007)?; // x7 <- data0
+
+        let abstractcs: Abstractcs = self.dmi.read_dmi_reg()?;
+        log::trace!("{:?}", abstractcs);
+        if abstractcs.busy() {
+            return Err(Error::AbstractCommandError(AbstractcsCmdErr::Busy)); //resue busy
+        }
+        if abstractcs.cmderr() != 0 {
+            AbstractcsCmdErr::try_from_cmderr(abstractcs.cmderr() as _)?;
+        }
+        Ok(())
+    }
 
     /// Read register value
     /// CSR: 0x0000 - 0x0fff
@@ -221,6 +265,294 @@ impl<'a, D: DebugModuleInterface> Algorigthm<'a, D> {
             let val = self.read_reg(*addr)?;
             log::info!("{}: 0x{:08x}", name, val);
         }
+
+        Ok(())
+    }
+
+    pub fn modify_mem32<F>(&mut self, addr: u32, f: F) -> Result<()>
+    where
+        F: FnOnce(u32) -> u32,
+    {
+        let data = self.read_mem32(addr)?;
+        let data = f(data);
+        self.write_mem32(addr, data)?;
+        Ok(())
+    }
+
+    pub fn wait_mem32<F>(&mut self, addr: u32, until: F) -> Result<u32>
+    where
+        F: Fn(u32) -> bool,
+    {
+        loop {
+            let data = self.read_mem32(addr)?;
+            if until(data) {
+                return Ok(data);
+            }
+            thread::sleep(Duration::from_millis(1));
+        }
+    }
+
+    pub fn read_memory(&mut self, addr: u32, len: u32) -> Result<Vec<u8>> {
+        if len % 4 != 0 {
+            return Err(Error::Custom("len must be 4 bytes aligned".to_string()));
+        }
+
+        let mut ret = Vec::with_capacity(len as usize);
+        for i in 0..len / 4 {
+            let data = self.read_mem32(addr + i * 4)?;
+            ret.extend_from_slice(&data.to_le_bytes());
+        }
+        Ok(ret)
+    }
+
+    fn lock_flash(&mut self) -> Result<()> {
+        const FLASH_CTLR: u32 = 0x40022010;
+
+        self.modify_mem32(FLASH_CTLR, |r| r | 0x00008080)?;
+        Ok(())
+    }
+
+    /// unlock FLASH LOCK and FLOCK
+    fn unlock_flash(&mut self) -> Result<()> {
+        const FLASH_CTLR: u32 = 0x40022010;
+        const FLASH_KEYR: u32 = 0x40022004;
+        const FLASH_MODEKEYR: u32 = 0x40022024;
+        const KEY1: u32 = 0x45670123;
+        const KEY2: u32 = 0xCDEF89AB;
+
+        let flash_ctlr = self.read_mem32(FLASH_CTLR)?;
+        log::debug!("flash_ctlr: 0x{:08x}", flash_ctlr);
+        // Test LOCK, FLOCK bits
+        if flash_ctlr & 0x00008080 == 0 {
+            // already unlocked
+            return Ok(());
+        }
+        // unlock LOCK
+        self.write_mem32(FLASH_KEYR, KEY1)?;
+        self.write_mem32(FLASH_KEYR, KEY2)?;
+
+        // unlock FLOCK
+        self.write_mem32(FLASH_MODEKEYR, KEY1)?;
+        self.write_mem32(FLASH_MODEKEYR, KEY2)?;
+
+        let flash_ctlr = self.read_mem32(FLASH_CTLR)?;
+        log::debug!("flash_ctlr: 0x{:08x}", flash_ctlr);
+
+        Ok(())
+    }
+
+    /// Erase by 256 bytes page
+    /// address must be 256 bytes aligned
+    pub fn fast_erase(&mut self, address: u32) -> Result<()> {
+        // require unlock
+        self.unlock_flash()?;
+
+        const FLASH_STATR: u32 = 0x4002200C;
+        const BUSY_MASK: u32 = 0x00000001;
+        const START_MASK: u32 = 1 << 6;
+        // const EOP_MASK: u32 = 1 << 5;
+        const WPROTECT_ERR_MASK: u32 = 1 << 4;
+
+        const FLASH_ADDR: u32 = 0x40022014;
+        const FLASH_CTLR: u32 = 0x40022010;
+
+        const PAGE_ERASE_MASK: u32 = 1 << 17;
+
+        if address & 0xff != 0 {
+            return Err(Error::Custom(
+                "address must be 256 bytes aligned".to_string(),
+            ));
+        }
+
+        let statr = self.read_mem32(FLASH_STATR)?;
+        // check if busy
+        if statr & BUSY_MASK != 0 {
+            return Err(Error::Custom("flash busy".to_string()));
+        }
+
+        self.modify_mem32(FLASH_CTLR, |r| r | PAGE_ERASE_MASK)?;
+
+        self.write_mem32(FLASH_ADDR, address)?;
+
+        self.modify_mem32(FLASH_CTLR, |r| r | START_MASK)?;
+
+        loop {
+            let statr = self.read_mem32(FLASH_STATR)?;
+            // check if busy
+            if statr & BUSY_MASK != 0 {
+                thread::sleep(Duration::from_millis(1));
+            } else {
+                if statr & WPROTECT_ERR_MASK != 0 {
+                    return Err(Error::Custom("flash write protect error".to_string()));
+                }
+                self.write_mem32(FLASH_STATR, statr)?; // write 1 to clear EOP
+
+                break;
+            }
+        }
+        // read 1 word to verify
+        let word = self.read_mem32(address)?;
+        println!("=> {:08x}", word);
+
+        // end erase, disable page erase
+        self.modify_mem32(FLASH_CTLR, |r| r & (!PAGE_ERASE_MASK))?;
+
+        self.lock_flash()?;
+
+        Ok(())
+    }
+
+    pub fn fast_erase_32k(&mut self, address: u32) -> Result<()> {
+        // require unlock
+        self.unlock_flash()?;
+
+        const FLASH_STATR: u32 = 0x4002200C;
+        const BUSY_MASK: u32 = 0x00000001;
+        const START_MASK: u32 = 1 << 6;
+        const WPROTECT_ERR_MASK: u32 = 1 << 4;
+
+        const FLASH_ADDR: u32 = 0x40022014;
+        const FLASH_CTLR: u32 = 0x40022010;
+
+        const BLOCK_ERASE_32K_MASK: u32 = 1 << 18;
+
+        if address & 0x7fff != 0 {
+            return Err(Error::Custom(
+                "address must be 32k bytes aligned".to_string(),
+            ));
+        }
+
+        let statr = self.read_mem32(FLASH_STATR)?;
+        // check if busy
+        if statr & BUSY_MASK != 0 {
+            return Err(Error::Custom("flash busy".to_string()));
+        }
+
+        self.modify_mem32(FLASH_CTLR, |r| r | BLOCK_ERASE_32K_MASK)?;
+
+        self.write_mem32(FLASH_ADDR, address)?;
+
+        self.modify_mem32(FLASH_CTLR, |r| r | START_MASK)?;
+
+        loop {
+            let statr = self.read_mem32(FLASH_STATR)?;
+            // check if busy
+            if statr & BUSY_MASK != 0 {
+                thread::sleep(Duration::from_millis(1));
+            } else {
+                if statr & WPROTECT_ERR_MASK != 0 {
+                    return Err(Error::Custom("flash write protect error".to_string()));
+                }
+                self.write_mem32(FLASH_STATR, statr)?; // write 1 to clear EOP
+
+                break;
+            }
+        }
+        // read 1 word to verify
+        let word = self.read_mem32(address)?;
+        println!("=> {:08x}", word);
+
+        // end erase
+        // disable page erase
+        self.modify_mem32(FLASH_CTLR, |r| r & (!BLOCK_ERASE_32K_MASK))?;
+
+        self.lock_flash()?;
+
+        Ok(())
+    }
+
+    pub fn erase_all(&mut self) -> Result<()> {
+        const FLASH_STATR: u32 = 0x4002200C;
+        const BUSY_MASK: u32 = 0x00000001;
+
+        const FLASH_CTLR: u32 = 0x40022010;
+        const MASS_ERASE_MASK: u32 = 1 << 2; // MER
+        const START_MASK: u32 = 1 << 6;
+
+        self.unlock_flash()?;
+
+        self.modify_mem32(FLASH_CTLR, |r| r | MASS_ERASE_MASK)?;
+
+        self.modify_mem32(FLASH_CTLR, |r| r | START_MASK)?;
+
+        let statr = self.wait_mem32(FLASH_STATR, |r| r & BUSY_MASK == 0)?;
+        self.write_mem32(FLASH_STATR, statr)?; // write 1 to clear EOP
+
+        // clear MER
+        self.modify_mem32(FLASH_CTLR, |r| r & (!MASS_ERASE_MASK))?;
+
+        Ok(())
+    }
+
+    /// Program bytes.
+    ///
+    /// # Arguments
+    ///
+    /// * `address` - The start address of the flash page to program.
+    /// * `data` - The data to be written to the page.
+    ///
+    /// The page must be erased first
+    pub fn program_page(&mut self, address: u32, data: &[u8]) -> Result<()> {
+        // require unlock
+        self.unlock_flash()?;
+
+        const FLASH_STATR: u32 = 0x4002200C;
+        const BUSY_MASK: u32 = 0x00000001;
+        const WRITE_BUSY_MASK: u32 = 1 << 1;
+        const WPROTECT_ERR_MASK: u32 = 1 << 4;
+
+        const FLASH_CTLR: u32 = 0x40022010;
+        const PAGE_START_MASK: u32 = 1 << 21; // start page program
+        const PAGE_PROG_MASK: u32 = 1 << 16; //
+
+        if address & 0xff != 0 {
+            return Err(Error::Custom(
+                "address must be 256 bytes aligned".to_string(),
+            ));
+        }
+
+        // check if busy
+        let statr = self.read_mem32(FLASH_STATR)?;
+        if statr & BUSY_MASK != 0 {
+            return Err(Error::Custom("flash busy".to_string()));
+        }
+
+        //let ctlr = self.read_mem32(FLASH_CTLR)?;
+        //let ctlr = ctlr | PAGE_PROG_MASK;
+        //self.write_mem32(FLASH_CTLR, ctlr)?;
+        self.modify_mem32(FLASH_CTLR, |r| r | PAGE_PROG_MASK)?;
+
+        for (i, word) in data.chunks(4).enumerate() {
+            let word = u32::from_le_bytes(word.try_into().unwrap());
+            self.write_mem32(address + (i as u32 * 4), word)?;
+
+            // write busy wait
+            self.wait_mem32(FLASH_STATR, |r| r & WRITE_BUSY_MASK == 0)?;
+        }
+
+        // start fast page program
+        self.modify_mem32(FLASH_CTLR, |r| r | PAGE_START_MASK)?;
+
+        // busy wait
+        let statr = self.wait_mem32(FLASH_STATR, |r| r & BUSY_MASK == 0)?;
+
+        self.write_mem32(FLASH_STATR, statr)?; // write 1 to clear EOP
+        if statr & WPROTECT_ERR_MASK != 0 {
+            return Err(Error::Custom("flash write protect error".to_string()));
+        }
+
+        // verify
+        // read 1 word to verify
+        //let word = self.read_mem32(address)?;
+        //println!("=> {:08x}", word);
+
+        // end program, clear PAGE_PROG
+        //let ctlr = self.read_mem32(FLASH_CTLR)?;
+        //let ctlr = ctlr & (!PAGE_PROG_MASK); // disable page erase
+        //self.write_mem32(FLASH_CTLR, ctlr)?;
+        self.modify_mem32(FLASH_CTLR, |r| r & (!PAGE_PROG_MASK))?;
+
+        self.lock_flash()?;
 
         Ok(())
     }